+++ cypress/e2e/form-level/form-level.cy.js
@@ -156,7 +156,7 @@ describe('Home page', () => {
       cy.uploadFile('sample.STU3.json');
       cy.get('#title').should('have.value', 'Sample STU3 form');
       cy.contains('button.dropdown-toggle', 'Export').click();
-      cy.contains('button.dropdown-item', 'Export to file in R4 format').click();
+      cy.contains('button.dropdown-item', 'Export to file in FHIR R4 format').click();
       cy.readFile('cypress/downloads/Sample-STU3-form.R4.json').then((json) => {
         cy.contains('#resizableMiddle .navbar button', 'Preview').scrollIntoView().click();
         cy.contains('.mat-mdc-tab-labels span', 'View Questionnaire JSON').scrollIntoView().click();
@@ -188,7 +188,7 @@ describe('Home page', () => {
       cy.get('@firstRow').find('[id^="code.0.code_"]').should('have.value', '8358-4');
 
       cy.contains('button.dropdown-toggle', 'Export').click();
-      cy.contains('button.dropdown-item', 'Export to file in STU3 format').click();
+      cy.contains('button.dropdown-item', 'Export to file in FHIR STU3 format').click();
       cy.readFile('cypress/downloads/Sample-R4-form.STU3.json').then((json) => {
         cy.contains('#resizableMiddle .navbar button', 'Preview').scrollIntoView().click();
         cy.contains('.mat-mdc-tab-labels span', 'View Questionnaire JSON').scrollIntoView().click();
@@ -206,8 +206,8 @@ describe('Home page', () => {
       cy.uploadFile('sample.R4.json');
       cy.get('#title').should('have.value', 'Sample R4 form');
       cy.contains('button.dropdown-toggle', 'Export').click();
-      cy.contains('button.dropdown-item', 'Export to file in LHC-FORMS format').click();
-      cy.readFile('cypress/downloads/Sample-R4-form.LHC-FORMS.json').then((json) => {
+      cy.contains('button.dropdown-item', 'Export to file in LHC-Forms internal (and volatile) format').click();
+      cy.readFile('cypress/downloads/Sample-R4-form.LHC-Forms.json').then((json) => {
         cy.contains('#resizableMiddle .navbar button', 'Preview').scrollIntoView().click();
         cy.contains('.mat-mdc-tab-labels span', 'View Questionnaire JSON').scrollIntoView().click();
         cy.contains('.preview-json-tabs .mat-mdc-tab-labels span', 'R4').scrollIntoView().click();