+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project follows [Semantic Versioning](http://semver.org/).
 
 ## [9.4.2] 2024-05-08
 ### Added
-- Menu option to download LHC-FORMS format.
+- Menu option to download LHC-Forms format.
 
 ## [9.4.1] 2024-04-29
 ### Added